pyproject.toml
[tool.poetry]
name = "glap-io"
version = "1.0.0"
description = ""
authors = ["Pandapip1 <admin@pandapip1.tk>"]

[tool.poetry.dependencies]
python = "^3.7"
pymunk = "^5.6"
profanity-check = "^1.0"
tornado = "^6.0"
mysql-connector = "^2.2"
git-client = "^0.2.1"
gmpy = "^1.17"
orjson = "^3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
